LayoutTests/media/media-source/media-source-append-buffer-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(initSegment))
EXPECTED (sourceBuffer.updating == 'true') OK
EVENT(updatestart)
EVENT(update)
EVENT(updateend)
EXPECTED (sourceBuffer.updating == 'false') OK
END OF TEST
